refactor: Complete server.py refactoring - 91% reduction in complexity
Successfully refactored 1441-line monolithic server.py into modular architecture:

**Before:**
- server.py: 1441 lines (monolithic, hard to maintain)

**After:**
- server.py: 124 lines (91% reduction, clean and focused)
- server_old.py: Backup of original implementation

**New Architecture:**
✅ Uses HandlerRouter for tool dispatching
✅ Uses ToolRegistry for centralized tool definitions
✅ All 23 tools routed through modular handlers
✅ Clean separation: server handles MCP protocol, handlers handle business logic

**Key Changes:**
1. Removed all handler methods (_handle_*) - now in handler classes
2. Simplified list_tools() - returns ToolRegistry.get_all_tools()
3. Simplified call_tool() - routes via HandlerRouter.route_tool_call()
4. Preserved server context for handlers (logger, config, engines, parsers)
5. Code formatted with black for consistency

**Benefits Achieved:**
- 🎯 Single Responsibility: Server only manages MCP protocol
- 📦 Modularity: Each handler focused on specific domain
- 🧪 Testability: Handlers can be tested independently
- 🔧 Maintainability: Easy to locate and modify tool logic
- 🚀 Extensibility: Add new tools by creating handlers

**Testing Status:**
- Code formatted and linted successfully
- Ready for integration testing

This completes Phase 4 of the refactoring plan.
Next: Run tests and validate all 23 tools still function correctly.
