+++ src/mcp_server/handlers/thinking_handler.py
